backend/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-multipart>=0.0.20
python-jose[cryptography]>=3.5.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.39
pymysql>=1.1.1
alembic>=1.16.4
pillow>=10.4.0
httpx>=0.28.1
openai>=1.99.9
redis>=6.4.0
celery>=5.5.3
